+++ batch.go
@@ -10,10 +10,10 @@ import (
 // around batch that flushes batch's data to disk
 // as soon as the configurable limit is reached.
 type BatchWithFlusher struct {
+	mtx   sync.Mutex
 	db    dbm.DB    // This is only used to create new batch
 	batch dbm.Batch // Batched writing buffer.
 
-	mtx            sync.Mutex
 	flushThreshold int // The threshold to flush the batch to disk.
 }
 
@@ -57,13 +57,11 @@ func (b *BatchWithFlusher) Set(key, value []byte) error {
 		return err
 	}
 	if batchSizeAfter > b.flushThreshold {
-		if err := b.batch.Write(); err != nil {
+		b.mtx.Unlock()
+		if err := b.Write(); err != nil {
 			return err
 		}
-		if err := b.batch.Close(); err != nil {
-			return err
-		}
-		b.batch = b.db.NewBatchWithSize(b.flushThreshold)
+		b.mtx.Lock()
 	}
 	return b.batch.Set(key, value)
 }
@@ -81,26 +79,47 @@ func (b *BatchWithFlusher) Delete(key []byte) error {
 		return err
 	}
 	if batchSizeAfter > b.flushThreshold {
-		if err := b.batch.Write(); err != nil {
+		b.mtx.Unlock()
+		if err := b.Write(); err != nil {
 			return err
 		}
-		if err := b.batch.Close(); err != nil {
-			return err
-		}
-		b.batch = b.db.NewBatchWithSize(b.flushThreshold)
+		b.mtx.Lock()
 	}
 	return b.batch.Delete(key)
 }
 
 func (b *BatchWithFlusher) Write() error {
-	return b.batch.Write()
+	b.mtx.Lock()
+	defer b.mtx.Unlock()
+
+	if err := b.batch.Write(); err != nil {
+		return err
+	}
+	if err := b.batch.Close(); err != nil {
+		return err
+	}
+	b.batch = b.db.NewBatchWithSize(b.flushThreshold)
+	return nil
 }
 
 func (b *BatchWithFlusher) WriteSync() error {
-	return b.batch.WriteSync()
+	b.mtx.Lock()
+	defer b.mtx.Unlock()
+
+	if err := b.batch.WriteSync(); err != nil {
+		return err
+	}
+	if err := b.batch.Close(); err != nil {
+		return err
+	}
+	b.batch = b.db.NewBatchWithSize(b.flushThreshold)
+	return nil
 }
 
 func (b *BatchWithFlusher) Close() error {
+	b.mtx.Lock()
+	defer b.mtx.Unlock()
+
 	return b.batch.Close()
 }